+++ xbmc/addons/AddonDatabase.cpp
@@ -1014,7 +1014,7 @@ bool CAddonDatabase::DisableAddon(const std::string& addonID, AddonDisabledReaso
 
     const std::string sql =
         PrepareSQL("UPDATE installed SET enabled=0, disabledReason=%d WHERE addonID='%s'",
-                   disabledReason, addonID.c_str());
+                   static_cast<int>(disabledReason), addonID.c_str());
     m_pDS->exec(sql);
     return true;
   }
@@ -1110,7 +1110,7 @@ bool CAddonDatabase::AddUpdateRuleForAddon(const std::string& addonID, AddonUpda
 
     std::string sql =
         PrepareSQL("INSERT INTO update_rules(id, addonID, updateRule) VALUES(NULL, '%s', %d)",
-                   addonID.c_str(), updateRule);
+                   addonID.c_str(), static_cast<int>(updateRule));
     m_pDS->exec(sql);
     return true;
   }
@@ -1139,7 +1139,7 @@ bool CAddonDatabase::RemoveUpdateRuleForAddon(const std::string& addonID,
 
     if (updateRule != AddonUpdateRule::ANY)
     {
-      sql += PrepareSQL(" AND updateRule = %d", updateRule);
+      sql += PrepareSQL(" AND updateRule = %d", static_cast<int>(updateRule));
     }
 
     m_pDS->exec(sql);
