+++ xbmc/addons/interfaces/gui/General.cpp
@@ -180,7 +180,7 @@ int Interface_GUIGeneral::get_video_resolution(KODI_HANDLE kodiBase)
     return -1;
   }
 
-  return CServiceBroker::GetWinSystem()->GetGfxContext().GetVideoResolution();
+  return (int)CServiceBroker::GetWinSystem()->GetGfxContext().GetVideoResolution();
 }
 //@}
