+++ xbmc/addons/kodi-dev-kit/include/kodi/addon-instance/Peripheral.h
@@ -684,7 +684,7 @@ private:
                                ->PerformDeviceScan(peripherals);
     if (err == PERIPHERAL_NO_ERROR)
     {
-      *peripheral_count = peripherals.size();
+      *peripheral_count = static_cast<unsigned int>(peripherals.size());
       kodi::addon::Peripherals::ToStructs(peripherals, scan_results);
     }
 
@@ -713,7 +713,7 @@ private:
                                ->GetEvents(peripheralEvents);
     if (err == PERIPHERAL_NO_ERROR)
     {
-      *event_count = peripheralEvents.size();
+      *event_count = static_cast<unsigned int>(peripheralEvents.size());
       kodi::addon::PeripheralEvents::ToStructs(peripheralEvents, events);
     }
 
@@ -817,7 +817,7 @@ private:
                                ->GetFeatures(addonJoystick, controller_id, featuresVector);
     if (err == PERIPHERAL_NO_ERROR)
     {
-      *feature_count = featuresVector.size();
+      *feature_count = static_cast<unsigned int>(featuresVector.size());
       kodi::addon::JoystickFeatures::ToStructs(featuresVector, features);
     }
 
@@ -869,7 +869,7 @@ private:
                                ->GetIgnoredPrimitives(addonJoystick, primitiveVector);
     if (err == PERIPHERAL_NO_ERROR)
     {
-      *primitive_count = primitiveVector.size();
+      *primitive_count = static_cast<unsigned int>(primitiveVector.size());
       kodi::addon::DriverPrimitives::ToStructs(primitiveVector, primitives);
     }
