+++ xbmc/cores/RetroPlayer/RetroPlayer.cpp
@@ -342,7 +342,7 @@ void CRetroPlayer::SeekPercentage(float fPercent /* = 0 */)
 
   uint64_t totalTime = GetTotalTime();
   if (totalTime != 0)
-    SeekTime(totalTime * fPercent / 100.0f);
+    SeekTime(static_cast<int64_t>(totalTime * fPercent / 100.0f));
 }
 
 float CRetroPlayer::GetCachePercentage() const
@@ -397,7 +397,7 @@ void CRetroPlayer::SetSpeed(float speed)
     else if (speed == 0.0f)
       m_callback.OnPlayBackPaused();
 
-    SetSpeedInternal(speed);
+    SetSpeedInternal(static_cast<double>(speed));
 
     if (speed == 0.0f)
     {
