+++ xbmc/cores/VideoPlayer/DVDDemuxers/DemuxMVC.cpp
@@ -19,34 +19,31 @@
 */
 
 #include "DemuxMVC.h"
-
 #include "DVDDemuxUtils.h"
 #include "DVDInputStreams/DVDInputStream.h"
-#include "cores/FFmpeg.h"
 #include "cores/VideoPlayer/Interface/TimingConstants.h"
+#include "cores/FFmpeg.h"
 #include "utils/log.h"
 
-extern "C"
-{
+extern "C" {
 #include "libavutil/opt.h"
 };
 
 #define MVC_SEEK_TIME_WINDOW 75000 // experimental value depends on seeking accurate
 
-static int mvc_file_read(void* h, uint8_t* buf, int size)
+static int mvc_file_read(void *h, uint8_t* buf, int size)
 {
   CDVDInputStream* pInputStream = static_cast<CDemuxMVC*>(h)->m_pInput;
   int s = pInputStream->Read(buf, size);
 
-  if (pInputStream->IsEOF())
-  {
-    return AVERROR_EOF;
+  if (pInputStream->IsEOF()) {
+	  return AVERROR_EOF;
   }
 
   return s;
 }
 
-static int64_t mvc_file_seek(void* h, int64_t pos, int whence)
+static int64_t mvc_file_seek(void *h, int64_t pos, int whence)
 {
   CDVDInputStream* pInputStream = static_cast<CDemuxMVC*>(h)->m_pInput;
   if (whence == AVSEEK_SIZE)
@@ -81,13 +78,12 @@ bool CDemuxMVC::Open(CDVDInputStream* pInput)
   if (blockSize > 1)
     bufferSize = blockSize;
   auto buffer = (unsigned char*)av_malloc(bufferSize);
-  m_ioContext =
-      avio_alloc_context(buffer, bufferSize, 0, this, mvc_file_read, nullptr, mvc_file_seek);
+  m_ioContext = avio_alloc_context(buffer, bufferSize, 0, this, mvc_file_read, nullptr, mvc_file_seek);
 
   m_pFormatContext = avformat_alloc_context();
   m_pFormatContext->pb = m_ioContext;
 
-  const AVInputFormat* format = av_find_input_format("mpegts");
+  const AVInputFormat *format = av_find_input_format("mpegts");
   ret = avformat_open_input(&m_pFormatContext, m_pInput->GetFileName().c_str(), format, nullptr);
   if (ret < 0)
   {
@@ -116,8 +112,8 @@ bool CDemuxMVC::Open(CDVDInputStream* pInput)
   CLog::Log(LOGDEBUG, "{}: MVC m2ts has {} streams", __FUNCTION__, m_pFormatContext->nb_streams);
   for (unsigned i = 0; i < m_pFormatContext->nb_streams; i++)
   {
-    if (m_pFormatContext->streams[i]->codecpar->codec_id == AV_CODEC_ID_H264_MVC &&
-        m_pFormatContext->streams[i]->codecpar->extradata_size > 0)
+    if (m_pFormatContext->streams[i]->codecpar->codec_id == AV_CODEC_ID_H264_MVC
+      && m_pFormatContext->streams[i]->codecpar->extradata_size > 0)
     {
       m_nStreamIndex = i;
       break;
@@ -185,16 +181,18 @@ DemuxPacket* CDemuxMVC::Read()
     }
     else
     {
-      AVStream* stream = m_pFormatContext->streams[pkt->stream_index];
+      AVStream *stream = m_pFormatContext->streams[pkt->stream_index];
       newPkt = CDVDDemuxUtils::AllocateDemuxPacket(pkt->size);
       if (pkt->data)
         memcpy(newPkt->pData, pkt->data, pkt->size);
       newPkt->iSize = pkt->size;
       newPkt->iStreamId = stream->id;
-      newPkt->dts = ConvertTimestamp(pkt->dts, stream->time_base.den, stream->time_base.num);
-      newPkt->pts = ConvertTimestamp(pkt->pts, stream->time_base.den, stream->time_base.num);
-      newPkt->duration =
-          DVD_SEC_TO_TIME((double)pkt->duration * stream->time_base.num / stream->time_base.den);
+      newPkt->dts =
+        ConvertTimestamp(pkt->dts, stream->time_base.den, stream->time_base.num);
+      newPkt->pts =
+        ConvertTimestamp(pkt->pts, stream->time_base.den, stream->time_base.num);
+      newPkt->duration = DVD_SEC_TO_TIME((double)pkt->duration * stream->time_base.num /
+                                          stream->time_base.den);
       break;
     }
   }
@@ -209,13 +207,11 @@ bool CDemuxMVC::SeekTime(double time, bool backwards, double* startpts)
     return false;
 
   AVRational time_base = m_pFormatContext->streams[m_nStreamIndex]->time_base;
-  int64_t seek_pts =
-      av_rescale(DVD_MSEC_TO_TIME(time), time_base.den, (int64_t)time_base.num * AV_TIME_BASE);
+  int64_t seek_pts = av_rescale(DVD_MSEC_TO_TIME(time), time_base.den, (int64_t)time_base.num * AV_TIME_BASE);
   int64_t starttime = 0;
 
-  if (m_pFormatContext->start_time != AV_NOPTS_VALUE)
-    starttime = av_rescale(m_pFormatContext->start_time, time_base.den,
-                           (int64_t)time_base.num * AV_TIME_BASE);
+  if (m_pFormatContext->start_time != (int64_t)AV_NOPTS_VALUE)
+    starttime = av_rescale(m_pFormatContext->start_time, time_base.den, (int64_t)time_base.num * AV_TIME_BASE);
   if (starttime != 0)
     seek_pts += starttime;
   if (seek_pts < MVC_SEEK_TIME_WINDOW)
@@ -233,8 +229,7 @@ std::string CDemuxMVC::GetFileName()
   return m_pInput->GetFileName();
 }
 
-AVStream* CDemuxMVC::GetAVStream() const
-{
+AVStream* CDemuxMVC::GetAVStream() const {
   return m_pFormatContext ? m_pFormatContext->streams[m_nStreamIndex] : nullptr;
 }
 
@@ -255,8 +250,7 @@ void CDemuxMVC::Dispose()
   m_nStreamIndex = -1;
 }
 
-double CDemuxMVC::ConvertTimestamp(int64_t pts, int den, int num) const
-{
+double CDemuxMVC::ConvertTimestamp(int64_t pts, int den, int num) const {
   if (pts == AV_NOPTS_VALUE)
     return DVD_NOPTS_VALUE;
 
@@ -265,7 +259,8 @@ double CDemuxMVC::ConvertTimestamp(int64_t pts, int den, int num) const
   double timestamp = (double)pts * num / den;
   double starttime = 0.0;
 
-  if (m_menu_type != MenuType::NATIVE && m_pFormatContext->start_time != AV_NOPTS_VALUE)
+  if (m_menu_type != MenuType::NATIVE &&
+      m_pFormatContext->start_time != static_cast<int64_t>(AV_NOPTS_VALUE))
   {
     starttime = static_cast<double>(m_start_time) / AV_TIME_BASE;
   }
