+++ xbmc/cores/VideoPlayer/DVDSubtitles/webvtt/WebVTTHandler.cpp
@@ -480,8 +480,9 @@ void CWebVTTHandler::DecodeLine(std::string line, std::vector<subtitleData>* sub
 
         auto colorInfo =
             std::find_if(m_CSSColors.begin(), m_CSSColors.end(),
-                         [&](const std::pair<std::string, UTILS::COLOR::ColorInfo>& item)
-                         { return StringUtils::CompareNoCase(item.first, colorName) == 0; });
+                         [&](const std::pair<std::string, UTILS::COLOR::ColorInfo>& item) {
+                           return StringUtils::CompareNoCase(item.first, colorName) == 0;
+                         });
         if (colorInfo != m_CSSColors.end())
         {
           const uint32_t color = UTILS::COLOR::ConvertToBGR(colorInfo->second.colorARGB);
@@ -763,7 +764,7 @@ void CWebVTTHandler::CalculateTextPosition(std::string& subtitleText)
   }
   else if (settings->positionAlign == WebvttAlign::RIGHT) // line-right
   {
-    marginLeft = cuePosPixel - cueSizePixel;
+    marginLeft = static_cast<int>(cuePosPixel - cueSizePixel);
     marginRight = static_cast<int>(VIEWPORT_WIDTH - cuePosPixel);
   }
   else if (settings->positionAlign == WebvttAlign::CENTER)
