+++ xbmc/cores/VideoPlayer/DVDSubtitles/webvtt/WebVTTISOHandler.cpp
@@ -28,7 +28,7 @@
 
 namespace
 {
-constexpr int defaultSubDuration = 20 * DVD_TIME_BASE;
+constexpr int defaultSubDuration = 20 * static_cast<int>(DVD_TIME_BASE);
 // VTTEmptyCueBox
 constexpr uint32_t ISO_BOX_TYPE_VTTE = StreamUtils::MakeFourCC('v', 't', 't', 'e');
 // VTTCueBox
@@ -42,6 +42,7 @@ constexpr uint32_t ISO_BOX_TYPE_STTG = StreamUtils::MakeFourCC('s', 't', 't', 'g
 // CuePayloadBox
 constexpr uint32_t ISO_BOX_TYPE_PAYL = StreamUtils::MakeFourCC('p', 'a', 'y', 'l');
 } // unnamed namespace
+
 
 void CWebVTTISOHandler::DecodeStream(const char* buffer,
                                      int bufferSize,
